Revert "a few shard utils"
