+++ corehq/form_processor/tests/utils.py
@@ -50,7 +50,7 @@ class FormProcessorTestUtils(object):
     @unit_testing_only
     def delete_all_sql_cases(cls, domain=None):
         logger.debug("Deleting all SQL cases for domain %s", domain)
-        cls._delete_all_sql_sharded_models(CommCareCaseSQL, domain)
+        cls._delete_all_sql_sharded_modesl(CommCareCaseSQL, domain)
 
     @staticmethod
     def delete_all_ledgers(domain=None):
@@ -97,7 +97,7 @@ class FormProcessorTestUtils(object):
     @unit_testing_only
     def delete_all_sql_forms(cls, domain=None):
         logger.debug("Deleting all SQL xforms for domain %s", domain)
-        cls._delete_all_sql_sharded_models(XFormInstanceSQL, domain)
+        cls._delete_all_sql_sharded_modesl(XFormInstanceSQL, domain)
 
     @classmethod
     @unit_testing_only
@@ -107,7 +107,7 @@ class FormProcessorTestUtils(object):
 
     @staticmethod
     @unit_testing_only
-    def _delete_all_sql_sharded_models(model_class, domain=None):
+    def _delete_all_sql_sharded_modesl(model_class, domain=None):
         assert issubclass(model_class, PartitionedModel)
         from corehq.sql_db.util import get_db_aliases_for_partitioned_query
         dbs = get_db_aliases_for_partitioned_query()
