+++ corehq/sql_db/tests/utils.py
@@ -1,7 +1,5 @@
 from __future__ import absolute_import
-from django.conf import settings
 from sqlalchemy.exc import ProgrammingError
-from corehq.sql_db.config import partition_config
 
 from corehq.sql_db.connections import connection_manager
 from corehq.util.decorators import ContextDecorator
@@ -33,25 +31,3 @@ class temporary_database(ContextDecorator):
         finally:
             conn.close()
             self.root_engine.dispose()
-
-
-class DefaultShardingTestConfigMixIn(object):
-    """
-    Mixin with test cases to ensure sharding is setup according to the expected defaults
-    """
-    db1 = 'p1'
-    db2 = 'p2'
-
-    def test_settings(self):
-        """
-        The tests in this class assume a certain partitioned setup to ensure the
-        partitioning is working properly, so this test makes sure those assumptions
-        are valid.
-        """
-
-        self.assertEqual(len(settings.PARTITION_DATABASE_CONFIG['shards']), 2)
-        self.assertIn(self.db1, settings.PARTITION_DATABASE_CONFIG['shards'])
-        self.assertIn(self.db2, settings.PARTITION_DATABASE_CONFIG['shards'])
-        self.assertEqual(settings.PARTITION_DATABASE_CONFIG['shards'][self.db1], [0, 1])
-        self.assertEqual(settings.PARTITION_DATABASE_CONFIG['shards'][self.db2], [2, 3])
-        self.assertEqual(set(partition_config.get_form_processing_dbs()), set([self.db1, self.db2]))
